[cxxmodules] Resolve possible symlinks in the path of rdict files.
When reading rdict files from the module we use the module name to form an
entry in a pending rdict map. It is later used when a library is loaded to
locate the corresponding rdict file.

This patch resolves a potential symlinks and allows ROOT to find its rdict
files. It fixes the cmssw C++ modules setup.