Avoid clone of constrained term when using bach mode
When using batch mode, cloning of PDf components is not supported.
However the cloning will be needed in case global observables are from the data. For the moment we do not support this case in batch mode.
In the future we will need to restore the global observable values in case it is taken from the data  or we allow the global observables will change after the fit.