[TMVA][SOFIE] Use new version of im2col
With this new version the output matrix is obtained in a row-major format.
This gives consecutive data read and data write from/to memory
and therefore is at least 3 times faster.

Furthermore,  using this new version there is no need to traspose
anymore the input data when using Gemm and we are also faster there.

The new function is implemented in SOFIE_Common